requirements.txt
attrs==25.1.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.27.1
numpy==2.0.2
packaging==24.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
streamlit
joblib
pandas
numpy
rdkit-pypi
scikit-learn
xgboost
altair
